backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
databases[sqlite]==0.8.0
aiofiles==23.2.1
aiohttp==3.9.1
pydantic==2.5.0
scikit-learn==1.3.2
scipy==1.11.4
# Обновляем до последних версий согласно документации
paddlepaddle>=2.6.0
paddleocr>=2.8.0

